+++ lib/core/network/model/user_format_statics_dto.dart
@@ -9,7 +9,7 @@ part 'user_format_statics_dto.g.dart';
 class UserFormatStaticsDto with _$UserFormatStaticsDto {
   const factory UserFormatStaticsDto({
     @JsonKey(name: 'count') int? count,
-    @JsonKey(name: 'meanScore') int? meanScore,
+    @JsonKey(name: 'meanScore') double? meanScore,
     @JsonKey(name: 'minutesWatched') int? minutesWatched,
     @JsonKey(name: 'chaptersRead') int? chaptersRead,
     @Default([]) @JsonKey(name: 'mediaIds') List<int> mediaIds,
