+++ lib/core/network/model/user_genres_statics_dto.dart
@@ -8,7 +8,7 @@ part 'user_genres_statics_dto.g.dart';
 class UserGenreStaticsDto with _$UserGenreStaticsDto {
   const factory UserGenreStaticsDto({
     @JsonKey(name: 'count') int? count,
-    @JsonKey(name: 'meanScore') int? meanScore,
+    @JsonKey(name: 'meanScore') double? meanScore,
     @JsonKey(name: 'minutesWatched') int? minutesWatched,
     @JsonKey(name: 'chaptersRead') int? chaptersRead,
     @Default([]) @JsonKey(name: 'mediaIds') List<int> mediaIds,
