+++ lib/core/network/model/user_staff_statics_dto.dart
@@ -9,7 +9,7 @@ part 'user_staff_statics_dto.g.dart';
 class UserStaffStaticsDto with _$UserStaffStaticsDto {
   const factory UserStaffStaticsDto({
     @JsonKey(name: 'count') int? count,
-    @JsonKey(name: 'meanScore') int? meanScore,
+    @JsonKey(name: 'meanScore') double? meanScore,
     @JsonKey(name: 'minutesWatched') int? minutesWatched,
     @JsonKey(name: 'chaptersRead') int? chaptersRead,
     @Default([]) @JsonKey(name: 'mediaIds') List<int> mediaIds,
