+++ lib/core/network/model/user_studio_statics_dto.dart
@@ -9,7 +9,7 @@ part 'user_studio_statics_dto.g.dart';
 class UserStudioStaticsDto with _$UserStudioStaticsDto {
   const factory UserStudioStaticsDto({
     @JsonKey(name: 'count') int? count,
-    @JsonKey(name: 'meanScore') int? meanScore,
+    @JsonKey(name: 'meanScore') double? meanScore,
     @JsonKey(name: 'minutesWatched') int? minutesWatched,
     @JsonKey(name: 'chaptersRead') int? chaptersRead,
     @Default([]) @JsonKey(name: 'mediaIds') List<int> mediaIds,
